power panic: re-use target temps from RAM
If a print has been saved to RAM such as during a filament runout,
do not overwrite these saved values if a sudden
power panic appears.

Additionally, change the saved types to be the same as power panic when saving to RAM:
- Bed target temperature is uint8_t (0 to 255) instead of float
- Extruder target temperature is uint16_t instead of float

Doing this change allows us to re-use the same global variables and
avoid creating local variables during power panic.

Change in memory:
Flash: -246 bytes
SRAM: -5 bytes

+++ Firmware/Marlin.h
@@ -303,8 +303,8 @@ extern uint8_t saved_printing_type;
 #define PRINTING_TYPE_USB 1
 #define PRINTING_TYPE_NONE 2
 
-extern float saved_extruder_temperature; //!< Active extruder temperature
-extern float saved_bed_temperature; //!< Bed temperature
+extern uint16_t saved_extruder_temperature; //!< Active extruder temperature
+extern uint8_t saved_bed_temperature; //!< Bed temperature
 extern uint8_t saved_fan_speed; //!< Print fan speed, ranges from 0 to 255
 extern uint16_t saved_feedrate2;
